requirements.txt
folium==0.12.1
requests==2.25.1
opencage==2.0.0
shodan==1.25.0
lxml==4.6.3
beautifulsoup4==4.10.0
dnspython==2.1.0
Pillow==8.4.0
